Move MangaCoverRatios to another file
